@@ -15,7 +15,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
+        "php": ">=8.1",
         "ext-json": "*",
         "ext-mysqli": "*",
         "activecollab/containeraccess": "^2.0",
@@ -24,8 +24,8 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.0",
-        "phpunit/phpunit": "~5.0",
-        "pimple/pimple": "~3.0.0"
+        "phpunit/phpunit": "~9.0",
+        "pimple/pimple": "~3.5.0"
     },
     "autoload": {
         "psr-4": {
